Create README.md
SUPERGULE-SLAM is an advanced framework that synergizes the capabilities of SUPERGULE and ORB SLAM to elevate the performance of monocular SLAM systems. This innovative approach is designed to integrate points matched by SUPERGULE into the ORB SLAM process, thereby enriching the point cloud with higher quality and quantity of feature points. This integration not only tackles the inherent challenges faced by single-camera SLAM systems but also significantly enhances their robustness and accuracy.